[vertical-scroll]: Do not propagate programmatic scroll from disabled to enabled frames.
If a 'vertical-scroll' is disabled for an <iframe>, then it should not
be able to affect the vertical scroll position. One way to do that is
using scripted scrolling by calling "element.scrollIntoView()".

To block such frames (whose feature's disabled), programmatic
recusive scroll calls are not forwarded to parent frame (unless the
parent frame has the feature disabled as well). In ther words, if a
given <iframe> is the first frame in a branch which does not have the
feature, then calls to programmatic scrollIntoView() from inside the
<iframe>:
  * Will not change the scroll offset in any of the ancestor frames,
  * Will work within the scope of <iframe>, i.e., elements become
visible inside <iframe>'s bounds.

Link to explainer/design document for "vertical-scroll":
https://docs.google.com/document/d/1qiWelnMlsOHuT_CQ0Zm_qEAf54HS5DhoIvEDHBlfqps/edit?usp=sharing

Bug: 611982
Change-Id: I0e06b399ad890e263128b997cfbb04eb3bdd1494